requirements.txt
numpy==1.15.4
scipy
pyyaml
matplotlib
scikit-image
tqdm
sgp4
pytz
julian
pygeodesy